Consistify HEA I beam type to allow resolving extra|shared parameters.
Use it for the default length, price of vendor eisen-fendt.
Useful for Bill of materials generators.
So far it had to be encoded in the `.blend` files but that is:
* not findable from the command line (binary format)
* not single source (sharing not possible)
* not maintainable (difficult batch changing of parameters in a binary)
* even more redundant or ambiguous if an assembly is made of more than one parts
that are not joined into a single object. (In which object or mesh ID to put
the extra parameters unless all are chained by the BOM generator?)